python-cli/requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
qdrant-client>=1.7.0
tqdm>=4.66.0
